% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{calculate_sample_size}
\alias{calculate_sample_size}
\title{Calculate Sample Size for Evaluating a Diagnostic Test}
\usage{
calculate_sample_size(sn, sp, p, w = 0.1, CI = 0.95)
}
\arguments{
\item{sn}{Sensitivity of the diagnostic test.}

\item{sp}{Specificity of the diagnostic test.}

\item{p}{Prevalence of the disease.}

\item{w}{Desired width of the confidence interval (default is 0.10).}

\item{CI}{Confidence interval level, either 0.95 or 0.9 (default is 0.95).}
}
\value{
A data frame containing the calculated sample sizes and input parameters.
}
\description{
This function calculates the sample size needed for evaluating a diagnostic test based on sensitivity, specificity, prevalence, and desired precision.
}
\examples{
calculate_sample_size()
}
